backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
chromadb
pypdf
tiktoken
pytest
httpx